Tryign to get the error working
At the moment I'm missing the fact that the number of points
in the ground truth actually changes over the number of
iterations. So I need it to calculate using the algorithm results
themselves instead of using the global fitter result. Requires
some refactoring across Joan's results package, but I think
it will be OK.